JAVA-2331: convert IT to unit test
