+++ core/src/main/resources/reference.conf
@@ -1613,7 +1613,8 @@ datastax-java-driver {
       }
 
       # The time after which the node level metrics will be evicted.
-      # The eviction will happen only if none of the enabled node-level metrics is updated for a given node for a given time.
+      # The eviction will happen only if none of the enabled node-level
+      # metrics is updated for a given node for a given time.
       # When the expire-after will elapse, all metrics for a given node are removed.
       #
       # Required: no (defaults to 1 hour)
